Refactor Deframer UTs
