+++ Svc/Deframer/CMakeLists.txt
@@ -39,6 +39,7 @@ set(UT_SOURCE_FILES
   "${CMAKE_CURRENT_LIST_DIR}/test/ut-fprime-protocol/SendBuffer.cpp"
   "${CMAKE_CURRENT_LIST_DIR}/test/ut-fprime-protocol/Tester.cpp"
   "${CMAKE_CURRENT_LIST_DIR}/test/ut-fprime-protocol/TestMain.cpp"
+  "${CMAKE_CURRENT_LIST_DIR}/test/ut-fprime-protocol/UplinkFrame.cpp"
 )
 set(UT_MOD_DEPS STest)
 register_fprime_ut(Svc_Deframer_fprime_protocol)
